test: refactor test util library
