+++ projects/internal/test-util/src/lib/noop-driver/noop-driver.module.spec.ts
@@ -1,0 +1,9 @@
+import { expectNgModuleToBeGuarded } from '../expect-ng-module-to-be-guarded';
+
+import { NoopDriverModule } from './noop-driver.module';
+
+describe(NoopDriverModule.name, () => {
+  it('is guarded against direct import', () => {
+    expectNgModuleToBeGuarded(NoopDriverModule);
+  });
+});
