+++ projects/internal/test-util/src/lib/noop-driver/noop-driver.module.ts
@@ -1,0 +1,21 @@
+import { ModuleWithProviders, NgModule } from '@angular/core';
+
+import { NoopDriverRootModule } from './noop-driver-root.module';
+
+/**
+ * Service module for `NoopDriver`.
+ *
+ * Use `NoopDriverModule.forRoot` to import.
+ */
+@NgModule()
+export class NoopDriverModule {
+  static forRoot(): ModuleWithProviders<NoopDriverRootModule> {
+    return {
+      ngModule: NoopDriverRootModule,
+    };
+  }
+
+  constructor() {
+    throw new Error('Do not import NoopDriverModule directly. Use NoopDriverModule.forRoot.');
+  }
+}
