+++ projects/internal/test-util/src/public_api.ts
@@ -3,5 +3,6 @@
  */
 
 export * from './lib/expect-ng-module-to-be-guarded';
-export * from './lib/noop-driver';
+export * from './lib/noop-driver/noop-driver.module';
+export * from './lib/noop-driver/noop-driver.service';
 export * from './lib/resolve-dependency';
